Refactored i18n plugin management to avoid warnings
