+++ tests/unit/index.js
@@ -3,8 +3,8 @@ import { config } from '@vue/test-utils'
 import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
-import { createI18n } from 'vue-i18n'
-import i18nMessages from '@/locales/en.json'
+// import { createI18n } from 'vue-i18n'
+// import i18nMessages from '@/locales/en.json'
 
 // Temporarily using the Vue migration build. Sets it to V3 mode to remove
 // a lot of deprecation warnings.
@@ -16,27 +16,29 @@ configureCompat({
 // are used for testing. Translation warnings are silenced, since some of the
 // resources used for testing are dummy and have no corresponding entry in the
 // language pack.
-const i18nInstance = createI18n({
-  legacy: false,
-  globalInjection: true,
-  locale: 'en',
-  fallbackLocale: 'en',
-  messages: { en: i18nMessages },
-  missingWarn: false,
-  fallbackWarn: false,
-  warnHtmlMessage: false
-})
+// const i18nInstance = createI18n({
+//   legacy: false,
+//   globalInjection: true,
+//   stopEffectScope: false,
+//   locale: 'en',
+//   fallbackLocale: 'en',
+//   messages: { en: i18nMessages },
+//   missingWarn: false,
+//   fallbackWarn: false,
+//   warnHtmlMessage: false
+// })
 
 // Setup Vuetify
 const vuetifyInstance = createVuetify({ components, directives })
 
 // Setup global plugins to be added to all mounted components by VTU
-config.global.plugins = [i18nInstance, vuetifyInstance]
+// config.global.plugins = [i18nInstance, vuetifyInstance]
+config.global.plugins = [vuetifyInstance]
 
-// config.global.mocks = {
-//   $t: tKey => tKey,
-//   $te: tKey => true
-// }
+config.global.mocks = {
+  $t: tKey => tKey,
+  $te: _ => true
+}
 
 // require all test files (files that ends with .spec.js)
 const testsContext = require.context('./specs', true, /\.spec\.js$/)
